+++ device/juniper/x86_64-juniper_qfx5200-r0/media_settings.json
