+++ device/juniper/x86_64-juniper_qfx5200-r0/plugins/qfx5200_eeprom_data.py
