+++ device/juniper/x86_64-juniper_qfx5210-r0/led_proc_init.soc
