+++ device/juniper/x86_64-juniper_qfx5210-r0/pmon_daemon_control.json
@@ -1,3 +1,4 @@
 {
+    "skip_thermalctld": true,
     "skip_ledd": true
 }
