+++ platform/broadcom/sonic-platform-modules-juniper/qfx5200/service/qfx5200-platform-init.service
