--- platform/broadcom/sonic-platform-modules-juniper/qfx5210/sonic_platform/__init__.py
@@ -1,1 +1,0 @@
-import platform
