+++ platform/broadcom/sonic-platform-modules-juniper/setup.py
@@ -8,9 +8,9 @@ os.listdir
 setup(
    name='sonic_platform',
    version='1.0',
-   description='Module to initialize Juniper QFX5210-64X platforms',
+   description='Module to initialize Juniper platforms',
    
    packages=['sonic_platform'],
-   package_dir={'sonic_platform': 'qfx5210/sonic_platform'},
+   package_dir={'sonic_platform': 'sonic_platform'},
 )
